third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDisplacementMapElement-dom-xChannelSelector-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'xChannelSelector' attribute of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.getAttribute('xChannelSelector') is "B"
PASS displacementMap.getAttribute('xChannelSelector') is "R"
PASS successfullyParsed is true

TEST COMPLETE
